rework user data
It's now a map of AbstractType -> (String -> String). This is because
substitution and user data don't work together well since substitutions
need to preserve and join it and some constraints would need to copy it
to other types on resolve.

This way just the final type (its name) is what you need to get user
data.