Allow "make clean" with project clean (#293)
Cleaning a project cleans the Makefile before calling "make clean". Therefore the Makefile is cleaned up before it can be called. This fixes that by overriding cleanOutput in the ProgramDslBuilderParticipant. Not that nice, however the only other way I see is to make the Makefile a resource that isn't generated by our language, which is even less nice.